+++ src/Ocelot/Request/Mapper/IRequestMapperExceptionConditions.cs
@@ -1,0 +1,9 @@
+﻿using Microsoft.AspNetCore.Http;
+
+namespace Ocelot.Request.Mapper
+{
+    public interface IRequestMapperExceptionConditions
+    {
+        bool PayloadTooLargeOnAnyHostedServer(HttpRequest request, Exception ex);
+    }
+}
