+++ src/Ocelot/Request/Mapper/RequestMapper.cs
@@ -1,35 +1,18 @@
-﻿using Microsoft.AspNetCore.Builder;
-using Microsoft.AspNetCore.Hosting.Server;
-using Microsoft.AspNetCore.Http;
+﻿using Microsoft.AspNetCore.Http;
 using Microsoft.AspNetCore.Http.Extensions;
-using Microsoft.AspNetCore.Server.HttpSys;
-using Microsoft.AspNetCore.Server.Kestrel.Core;
-using Microsoft.Extensions.Options;
 using Microsoft.Extensions.Primitives;
 using Ocelot.Configuration;
 using Ocelot.Responses;
-using System.Diagnostics;
-using System.Runtime.InteropServices;
 
 namespace Ocelot.Request.Mapper
 {
     public class RequestMapper : IRequestMapper
     {
         private readonly string[] _unsupportedHeaders = { "host" };
-        private readonly long? maxRequestBodySizeHttpsSys;
-        private readonly long? maxRequestBodySizeKerstrelServer;
-        private readonly long? maxRequestBodySizeIISServer;
-        private IServer _server;
-        public RequestMapper()
+        private readonly IRequestMapperExceptionConditions _requestMapperExceptionConditions;
+        public RequestMapper(IRequestMapperExceptionConditions requestMapperExceptionConditions)
         {
-        }
-
-        public RequestMapper(IOptions<HttpSysOptions> httpSysOptions, IOptions<KestrelServerOptions> kerstrelServerOptions, IOptions<IISServerOptions> iISOptions,IServer server)
-        {
-            maxRequestBodySizeHttpsSys = httpSysOptions.Value.MaxRequestBodySize.GetValueOrDefault();
-            maxRequestBodySizeKerstrelServer = kerstrelServerOptions.Value.Limits.MaxRequestBodySize;
-            maxRequestBodySizeIISServer = iISOptions.Value.MaxRequestBodySize;
-            _server = server;
+            _requestMapperExceptionConditions = requestMapperExceptionConditions;
         }
 
         public async Task<Response<HttpRequestMessage>> Map(HttpRequest request, DownstreamRoute downstreamRoute)
@@ -50,49 +33,13 @@ namespace Ocelot.Request.Mapper
             }
             catch (Exception ex)
             {
-                if (PayloadTooLargeOnAnyHostedServer(request, ex))
+                if (_requestMapperExceptionConditions.PayloadTooLargeOnAnyHostedServer(request, ex))
                 {
                     return new ErrorResponse<HttpRequestMessage>(new PayloadTooLargeError(ex));
                 }
 
                 return new ErrorResponse<HttpRequestMessage>(new UnmappableRequestError(ex));
             }
-        }
-
-        private bool PayloadTooLargeOnAnyHostedServer(HttpRequest request, Exception ex)
-        {
-            bool isHeavyPayload = false;
-
-            if(_server is KestrelServer)
-            {
-                isHeavyPayload = maxRequestBodySizeKerstrelServer < request.ContentLength;
-            }
-            else if (IsRunningInProcessIIS())
-            {
-               isHeavyPayload = maxRequestBodySizeIISServer < request.ContentLength;
-            }
-            else
-            {
-                isHeavyPayload = maxRequestBodySizeHttpsSys < request.ContentLength && RuntimeInformation.IsOSPlatform(OSPlatform.Windows);
-            }
-
-            return isHeavyPayload && ex is Microsoft.AspNetCore.Http.BadHttpRequestException;
-        }
-
-        /// <summary>
-        /// Check if this process is running on Windows in an in process instance in IIS.
-        /// </summary>
-        /// <returns>True if Windows and in an in process instance on IIS, false otherwise.</returns>
-        private static bool IsRunningInProcessIIS()
-        {
-            if (!RuntimeInformation.IsOSPlatform(OSPlatform.Windows))
-            {
-                return false;
-            }
-
-            string processName = Path.GetFileNameWithoutExtension(Process.GetCurrentProcess().ProcessName);
-            return processName.Contains("w3wp", StringComparison.OrdinalIgnoreCase) ||
-                processName.Contains("iisexpress", StringComparison.OrdinalIgnoreCase);
         }
 
         private static async Task<HttpContent> MapContent(HttpRequest request)
